+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: assistantnms_app
 description: Assistant for No Man's Sky
 homepage: https://nmsassistant.com
 # author: Kurt Lourens
-version: 2.16.1+317 # dart scripts\version_num_script.dart
+version: 2.16.2+318 # dart scripts\version_num_script.dart
 publish_to: 'none'
 
 environment:
@@ -97,7 +97,7 @@ flutter_icons: ## flutter pub run flutter_launcher_icons:main
 msix_config: # https://pub.dev/documentation/msix/latest/
   display_name: Assistant for No Man's Sky
   publisher_display_name: AssistantApps
-  msix_version: 2.16.1.0
+  msix_version: 2.16.2.0
   identity_name: notSureIfSecret
   publisher: notSureIfSecret
   logo_path: './assets/images/fullSizeIcon.png'
